requirements.txt
flet>=0.84.0
pillow>=10.0.0